update-manual final, aliases

+++ scripts/clear-cache.sh
@@ -2,8 +2,6 @@
 
 # Script for clearing cache, distro-agnostic, can be used whenever
 
-# NOTE: This removes all NeoVim cache, which means that you'll reinstall plugins on launch
-
 sudo rm -rf ~/.cache/thumbnails
 sudo rm -rf ~/.cache/nsxiv
 sudo rm -rf ~/.cache/qimgv/thumbnails
@@ -15,9 +13,20 @@ sudo rm -rf ~/.var/app/dev.vencord.Vesktop/config/vesktop/sessionData/Cache/Cach
 sudo rm -rf "/home/stankovictab/.var/app/dev.vencord.Vesktop/config/vesktop/sessionData/Code Cache"
 sudo rm -rf ~/.var/app/com.discordapp.Discord/config/discord/Cache/Cache_Data
 
-sudo rm -rf ~/.cache/nvim
-sudo rm -rf ~/.local/share/nvim
-sudo rm -rf ~/.local/state/nvim
+# NOTE: This removes all NeoVim cache, which means that you'll reinstall plugins on launch
+read -p "Do you want to clear NeoVim cache? [Y/n] " response
+# Convert response to lowercase and handle empty input (Enter) as Y
+response=${response,,} # convert to lowercase
+response=${response:-y} # default to y if empty
+if [[ $response =~ ^[yY]$ ]]; then
+    echo "Clearing NeoVim cache..."
+    sudo rm -rf ~/.cache/nvim
+    sudo rm -rf ~/.local/share/nvim
+    sudo rm -rf ~/.local/state/nvim
+    echo "NeoVim cache cleared successfully!"
+else
+    echo "Neovim cache not cleared."
+fi
 
 sudo rm -rf ~/.config/obsidian/GPUCache # Fixes a lot of graphical issues with Obsidian
